requirements.txt
Flask==2.2.3
gunicorn==20.1.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.4
python-dotenv==0.21.1
Flask-Cors==3.0.10
connexion[swagger-ui]